added permissions to instructorregistration

+++ gymregistration/src/main/java/ca/mcgill/ecse321/gymregistration/service/InstructorRegistrationService.java
@@ -9,13 +9,15 @@ import org.springframework.stereotype.Service;
 
 import ca.mcgill.ecse321.gymregistration.dao.InstructorRegistrationRepository;
 import ca.mcgill.ecse321.gymregistration.dao.InstructorRepository;
+import ca.mcgill.ecse321.gymregistration.dao.OwnerRepository;
 import ca.mcgill.ecse321.gymregistration.dao.SessionRepository;
+import ca.mcgill.ecse321.gymregistration.model.GymUser;
 import ca.mcgill.ecse321.gymregistration.model.Instructor;
 import ca.mcgill.ecse321.gymregistration.model.InstructorRegistration;
+import ca.mcgill.ecse321.gymregistration.model.Owner;
 import ca.mcgill.ecse321.gymregistration.model.Session;
 import ca.mcgill.ecse321.gymregistration.service.exception.GRSException;
 import jakarta.transaction.Transactional;
-
 
 @Service
 public class InstructorRegistrationService {
@@ -26,6 +28,8 @@ public class InstructorRegistrationService {
     private SessionRepository sessionRepository;
     @Autowired
     private InstructorRegistrationRepository instructorRegistrationRepository;
+    @Autowired
+    private OwnerRepository ownerRepository;
 
     /**
      * Create a new instructor registeration to add an instructor to a class
@@ -39,19 +43,17 @@ public class InstructorRegistrationService {
         Instructor instructor = instructorRepository.findInstructorById(instructorId);
         Session session = sessionRepository.findSessionById(sessionId);
         if (instructor == null) {
-            throw new GRSException(HttpStatus.UNAUTHORIZED, "Instructor not found");
+            throw new GRSException(HttpStatus.NOT_FOUND, "Instructor not found");
         }
         if (session == null)
-            throw new GRSException(HttpStatus.UNAUTHORIZED, "Session not found");
+            throw new GRSException(HttpStatus.NOT_FOUND, "Session not found");
         InstructorRegistration instructorRegistration = instructorRegistrationRepository
                 .findInstructorRegistrationByInstructor_idAndSession_id(instructorId, sessionId);
         if (instructorRegistration != null)
-            throw new GRSException(HttpStatus.UNAUTHORIZED, "already registered");
+            throw new GRSException(HttpStatus.BAD_REQUEST, "already registered");
 
         instructorRegistration = new InstructorRegistration(null, instructor, session);
-
         instructorRegistrationRepository.save(instructorRegistration);
-
         return instructorRegistration;
     }
 
@@ -63,22 +65,25 @@ public class InstructorRegistrationService {
      * @throws GRSException not able to remove instructor
      */
     @Transactional
-
-    public void removeInstructorFromClass(int sessionId, int instructorId)
-    {
-        List<InstructorRegistration> instructorRegistrations = instructorRegistrationRepository.findInstructorRegistrationsBySession_id(sessionId);
-        if(instructorRegistrations.size() <2)
-            throw new GRSException(HttpStatus.UNAUTHORIZED, "not enough instructors registered");
-        for(InstructorRegistration r: instructorRegistrations )
-        {
-
-            if(r.getInstructor().getId() == instructorId)
+    public void removeInstructorFromClass(int sessionId, int instructorId, int gymuserId) {
+        List<InstructorRegistration> instructorRegistrations = instructorRegistrationRepository
+                .findInstructorRegistrationsBySession_id(sessionId);
+        if (instructorRegistrations.size() < 2)
+            throw new GRSException(HttpStatus.BAD_REQUEST, "not enough instructors registered");
+        GymUser gymuser = instructorRepository.findInstructorById(gymuserId);
+        if(gymuser ==null)
+            gymuser = ownerRepository.findOwnerById(gymuserId);
+        
+        if( gymuser == null || gymuser instanceof Owner == false || instructorRegistrationRepository.findInstructorRegistrationByInstructor_idAndSession_id(gymuser.getId(), sessionId)==null)
+            throw new GRSException(HttpStatus.UNAUTHORIZED, "You don't have permission to remove this instructor");
+        for (InstructorRegistration r : instructorRegistrations) {
+            if (r.getInstructor().getId() == instructorId)
                 instructorRegistrationRepository.delete(r);
             return;
         }
-        throw new GRSException(HttpStatus.UNAUTHORIZED, "instructor not teaching course");
+        throw new GRSException(HttpStatus.BAD_REQUEST, "instructor not teaching course");
+    }
 
-    }    
     /**
      * return an instructor registration
      * 
@@ -101,7 +106,7 @@ public class InstructorRegistrationService {
                 .findInstructorRegistrationByInstructor_idAndSession_id(instructor.getId().intValue(), session.getId());
 
         if (instructorRegistration == null)
-            throw new GRSException(HttpStatus.NOT_FOUND, "Instructor not registered for this session");
+            throw new GRSException(HttpStatus.BAD_REQUEST, "Instructor not registered for this session");
 
         return instructorRegistration;
     }
@@ -110,7 +115,7 @@ public class InstructorRegistrationService {
         InstructorRegistration instructorRegistration = instructorRegistrationRepository
                 .findInstructorRegistrationById(id);
         if (instructorRegistration == null)
-            throw new GRSException(HttpStatus.NOT_FOUND, "Instructor not registered for this session");
+            throw new GRSException(HttpStatus.BAD_REQUEST, "Instructor not registered for this session");
         return instructorRegistration;
     }
 }
